Convert all globals to module imports, and more
Move most of background.js into new tgs.js to allow import access to other modules
background.js retains the top-level event listeners required by service workers

Convert to new V3 favicon URL structure
Remove default favicon cache from gsFavicon (hopefully temporary)
Add new V3 favicon path to web accessible

Add scripting permission to allow content script injection
Remove unneeded references to "window" since service worker can't access anyway
